+++ src/tools/fscankeywords.c
@@ -17,9 +17,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
-#ifdef USE_JSON
 #include <json-c/json.h>
-#endif
 #include "types.h"
 
 Bool fscankeywords(LPJfile *file,    /**< pointer to LPJ file */
@@ -31,69 +29,39 @@ Bool fscankeywords(LPJfile *file,    /**< pointer to LPJ file */
                    Verbosity verb    /**< verbosity level (NO_ERR,ERR,VERB) */
                   )                  /** \return TRUE on error */
 {
-  String line,token;
-  char *ptr;
   const char *str;
-  Bool rc;
-#ifdef USE_JSON
   int i;
   struct json_object *item;
-  if(file->isjson)
+  if(name==NULL)
+    item=file;
+  else
   {
-    if(name==NULL)
-      item=file->file.obj;
-    else
+    if(!json_object_object_get_ex(file,name,&item))
     {
-      if(!json_object_object_get_ex(file->file.obj,name,&item))
-      {
-        if(with_default)
-        {
-          if(verb)
-            fprintf(stderr,"WARNING027: Name '%s' for keyword not found, set to \"%s\".\n",
-                    name,array[*value]);
-          return FALSE;
-        }
-        else
-        {
-          if(verb)
-            fprintf(stderr,"ERROR225: Name '%s' for keyword not found.\n",name);
-          return TRUE;
-        }
-      }
-    }
-    if(json_object_get_type(item)==json_type_string)
-    {
-      str=json_object_get_string(item);
-      *value=findstr(str,array,size);
-      if(*value==NOT_FOUND)
+      if(with_default)
       {
         if(verb)
-        {
-          fprintf(stderr,"ERROR166: Keyword '%s' not defined for name '%s', must be",str,(name==NULL) ? "N/A" : name);
-          for(i=0;i<size;i++)
-          {
-            fprintf(stderr," \"%s\"",array[i]);
-            if(i!=size-1)
-              fputc(',',stderr);
-          }
-          fputs(".\n",stderr);
-        }
+          fprintf(stderr,"WARNING027: Name '%s' for keyword not found, set to \"%s\".\n",
+                  name,array[*value]);
+        return FALSE;
+      }
+      else
+      {
+        if(verb)
+          fprintf(stderr,"ERROR225: Name '%s' for keyword not found.\n",name);
         return TRUE;
       }
-      return FALSE;
     }
-    if(json_object_get_type(item)!=json_type_int)
-    {
-      if(verb)
-        fprintf(stderr,"ERROR226: Name '%s' not of type int.\n",(name==NULL) ? "N/A" : NULL);
-      return TRUE;
-    }
-    *value=json_object_get_int(item);
-    if(*value<0 || *value>=size)
+  }
+  if(json_object_get_type(item)==json_type_string)
+  {
+    str=json_object_get_string(item);
+    *value=findstr(str,array,size);
+    if(*value==NOT_FOUND)
     {
       if(verb)
       {
-        fprintf(stderr,"ERROR226: Invalid value %d for name '%s', must be in [0,%d] or",*value,(name==NULL) ?  "N/A" : name,size-1);
+        fprintf(stderr,"ERROR166: Keyword '%s' not defined for name '%s', must be",str,name);
         for(i=0;i<size;i++)
         {
           fprintf(stderr," \"%s\"",array[i]);
@@ -104,44 +72,31 @@ Bool fscankeywords(LPJfile *file,    /**< pointer to LPJ file */
       }
       return TRUE;
     }
-    if (verb >= VERB)
-      printf("\"%s\" : %d\n", name, *value);
     return FALSE;
   }
-#endif
-  rc=fscantoken(file->file.file,token);
-  if(!rc)
+  if(json_object_get_type(item)!=json_type_int)
   {
-    *value=(int)strtol(token,&ptr,10);
-    rc=*ptr!='\0';
-    if(*value<0 || *value>=size)
-    {
-      if(verb)
-        fprintf(stderr,"ERROR166: Invalid value %d for name '%s' in line %d of '%s'.\n",*value,name,getlinecount(),getfilename());
-      return TRUE;
-    }
+    if(verb)
+      fprintf(stderr,"ERROR226: Name '%s' not of type int.\n",name);
+    return TRUE;
   }
-  if(rc && verb)
+  *value=json_object_get_int(item);
+  if(*value<0 || *value>=size)
   {
-    fprintf(stderr,"ERROR101: Cannot read int '%s' in line %d of '%s', ",
-            name,getlinecount(),getfilename());
-    if(strlen(token)>0)
+    if(verb)
     {
-      fputs("read:\n",stderr);
-      if(fgets(line,STRING_LEN,file->file.file)!=NULL)
-        line[strlen(line)-1]='\0';
-      else
-        line[0]='\0';
-      fputs("          '",stderr);
-      fputprintable(stderr,token);
-      fprintf(stderr,"%s'\n           ",line);
-      frepeatch(stderr,'^',strlen(token));
-      fputc('\n',stderr);
+      fprintf(stderr,"ERROR226: Invalid value %d for name '%s', must be in [0,%d] or",*value,name,size-1);
+      for(i=0;i<size;i++)
+      {
+        fprintf(stderr," \"%s\"",array[i]);
+        if(i!=size-1)
+          fputc(',',stderr);
+      }
+      fputs(".\n",stderr);
     }
-    else
-      fputs("EOF reached.\n",stderr);
+    return TRUE;
   }
-  else if (verb >= VERB)
+  if (verb >= VERB)
     printf("\"%s\" : %d\n", name, *value);
-  return rc;
+  return FALSE;
 } /* of 'fscankeywords' */